requirements-test.txt
wheel
pyyaml
bottle
requests
pyftpdlib
beaker
psutil
flake8
WebTest
bcrypt
Pillow
vdf
pyudev
plyvel
pytest
pytidylib
requests-mock
pyfakefs
pyright
